app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
requests==2.31.0
pytz==2024.1
